refactor(图表): 去除G2PLOT
