+++ core/core-frontend/package.json
@@ -15,7 +15,6 @@
   },
   "dependencies": {
     "@antv/g2": "^5.3.2",
-    "@antv/g2plot": "^2.4.29",
     "@antv/l7": "^2.22.0",
     "@antv/l7plot": "^0.5.5",
     "@antv/s2": "^1.49.0",
