+++ core/core-frontend/src/views/chart/components/js/panel/charts/g2plot/bar/bar.ts
@@ -17,7 +17,7 @@ import {
   BAR_AXIS_TYPE,
   BAR_EDITOR_PROPERTY,
   BAR_EDITOR_PROPERTY_INNER
-} from '@/views/chart/components/js/panel/charts/bar/common'
+} from '@/views/chart/components/js/panel/charts/g2plot/bar/common'
 import {
   assembleOptionsDataForRoundAngle,
   configPlotTooltipEvent,
