+++ core/core-frontend/src/views/chart/components/js/panel/charts/g2plot/others/chart-mix.ts
@@ -15,7 +15,7 @@ import {
   getYAxisExt,
   setGradientColor,
   TOOLTIP_TPL
-} from '../../common/common_antv'
+} from '../../../common/common_antv'
 import { flow, hexColorToRGBA, parseJson } from '@/views/chart/components/js/util'
 import {
   cloneDeep,
