+++ core/core-frontend/src/views/chart/components/js/panel/charts/g2plot/others/funnel.ts
@@ -1,7 +1,7 @@
 import type { FunnelOptions, Funnel as G2Funnel } from '@antv/g2plot/esm/plots/funnel'
-import { G2PlotChartView, G2PlotDrawOptions } from '../../types/impl/g2plot'
+import { G2PlotChartView, G2PlotDrawOptions } from '../../../types/impl/g2plot'
 import { flow, parseJson, setUpSingleDimensionSeriesColor } from '@/views/chart/components/js/util'
-import { configPlotTooltipEvent, getPadding } from '../../common/common_antv'
+import { configPlotTooltipEvent, getPadding } from '../../../common/common_antv'
 import { useI18n } from '@/hooks/web/useI18n'
 import { Datum } from '@antv/g2plot/esm/types/common'
 import { valueFormatter } from '@/views/chart/components/js/formatter'
