+++ core/core-frontend/src/views/chart/components/js/panel/charts/g2plot/others/picture-group.ts
@@ -1,4 +1,4 @@
-import { AbstractChartView, ChartLibraryType, ChartRenderType } from '../../types'
+import { AbstractChartView, ChartLibraryType, ChartRenderType } from '../../../types'
 import { useI18n } from '@/hooks/web/useI18n'
 
 const { t } = useI18n()
