+++ core/core-frontend/src/views/chart/components/js/panel/charts/g2plot/others/radar.ts
@@ -1,8 +1,8 @@
 import type { RadarOptions, Radar as G2Radar } from '@antv/g2plot/esm/plots/radar'
-import { G2PlotChartView, G2PlotDrawOptions } from '../../types/impl/g2plot'
-import { flow, parseJson } from '../../../util'
-import { configPlotTooltipEvent } from '../../common/common_antv'
-import { valueFormatter } from '../../../formatter'
+import { G2PlotChartView, G2PlotDrawOptions } from '../../../types/impl/g2plot'
+import { flow, parseJson } from '../../../../util'
+import { configPlotTooltipEvent } from '../../../common/common_antv'
+import { valueFormatter } from '../../../../formatter'
 import type { Datum } from '@antv/g2plot/esm/types/common'
 import { useI18n } from '@/hooks/web/useI18n'
 import { DEFAULT_LABEL, DEFAULT_LEGEND_STYLE } from '@/views/chart/components/editor/util/chart'
