+++ core/core-frontend/src/views/chart/components/js/panel/charts/g2plot/others/treemap.ts
@@ -1,8 +1,8 @@
 import { TreemapOptions, Treemap as G2Treemap } from '@antv/g2plot/esm/plots/treemap'
-import { G2PlotChartView, G2PlotDrawOptions } from '../../types/impl/g2plot'
-import { flow, parseJson, setUpSingleDimensionSeriesColor } from '../../../util'
-import { getPadding, getTooltipSeriesTotalMap } from '../../common/common_antv'
-import { valueFormatter } from '../../../formatter'
+import { G2PlotChartView, G2PlotDrawOptions } from '../../../types/impl/g2plot'
+import { flow, parseJson, setUpSingleDimensionSeriesColor } from '../../../../util'
+import { getPadding, getTooltipSeriesTotalMap } from '../../../common/common_antv'
+import { valueFormatter } from '../../../../formatter'
 import { Label } from '@antv/g2plot/lib/types/label'
 import { Datum } from '@antv/g2plot/esm/types/common'
 import { useI18n } from '@/hooks/web/useI18n'
