+++ core/core-frontend/src/views/chart/components/js/panel/charts/g2plot/others/word-cloud.ts
